Enhancement: Source Tree Refactor
