+++ include/arch/core/or1k/cache.h
@@ -35,7 +35,8 @@
 
 #ifndef _ASM_FILE_
 
-	#include <arch/core/or1k/core.h>
+	#define __NEED_OR1K_REGS
+	#include <arch/core/or1k/regs.h>
 
 #endif /* _ASM_FILE_ */
