+++ include/arch/core/or1k/regs.h
@@ -440,4 +440,52 @@
 	#define OR1K_EXCEPTION_LOAD_GPR6(reg)  l.lwz reg, 0x74(r0)
 #endif
 
+#ifndef _ASM_FILE_
+
+	/**
+	 * @brief Reads from an specified Special-Purpose register.
+	 *
+	 * @param reg Register to be read.
+	 *
+	 * @returns Returns the value of the specified SPR.
+	 */
+	static inline unsigned or1k_mfspr(unsigned reg)
+	{
+		register unsigned ret
+			__asm__("r3") = (unsigned) 0;
+		register unsigned r5
+			__asm__("r5") = (unsigned) reg;
+
+		__asm__ __volatile__
+		(
+			"l.mfspr r3, r5, 0"
+			: "=r" (ret)
+			: "r" (r5)
+		);
+		return (ret);
+	}
+
+	/**
+	 * @brief Writes to an specified Special-Purpose register.
+	 *
+	 * @param reg Register to be written.
+	 * @param value Value to be written.
+	 */
+	static inline void or1k_mtspr(unsigned reg, unsigned value)
+	{
+		register unsigned r3
+			__asm__("r3") = (unsigned) reg;
+		register unsigned r5
+			__asm__("r5") = (unsigned) value;
+
+		__asm__ __volatile__
+		(
+			"l.mtspr r3, r5, 0\n"
+			:
+			: "r" (r3), "r" (r5)
+		);
+	}
+
+#endif
+
 #endif	/* ARCH_CORE_OR1K_REGS_H_ */
